Add files via upload
Cadaverplant (Sulfurimortia symbiotica)
This plant is very special, and very different from anything seen on earth. Named after it’s deadly berries and its symbiotic relationship with the ecosystem, it thrives in cold places. 
Its broad leaves have large surface areas to capture the sulphuric gas in the atmosphere. Its roots absorb the liquid ammonia in the soil, and through absorbing light, it would produce nitrogen compounds and hydrogen sulphide byproducts, this makes it extremely important for all life on the planet. Although a killer, it also allows life to thrive on its world. 
It secrets a chemical compound that attracts animals to graze upon its delicious but deadly berries. Once eaten, the sulfuric acid in the berries would begin to irritate the animal’s digestive system, it starts causing burns and lesions in the digestive tract. The plant has evolved to carefully control the amount of acid in the berries, this causes the animal not to die straight away but weaken it enough to make it wander a small distance away. This reduces the competition for resources the plants might face.
Some scientists think that the seeds of the plant start germinating straight away inside of an animal. This is possible due to the seeds protective shell. The plant will then start to feed off the nutrients inside of the animal. Eventually the animal would become lethargic and susceptible to predators. Some say that predators that have eaten the prey animal also ended up with seeds inside them, although this has not been empirically proven yet. The plant can also grow from animal droppings, should the animal be lucky enough to live that long.
Eventually the animal dies from organ failure or an inability to sustain itself due to the seeds consuming all other nutrients. Once the body starts decomposing the Cadaverplant will take root and start growing as a small shrub.